langchain/ice_breaker/pyproject.toml
[project]
name = "ice-breaker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.1.9",
    "langchain-openai>=0.3.29",
    "langchainhub>=0.1.21",
    "python-dotenv>=1.1.1",
]
